pyproject.toml
[project]
name = "sciedu-llm"
version = "0.1.0"
description = "An LLM service provider for NYCU's Cognition and Science Learning Laboratory"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openai>=2.15.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.40.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
]

[tool.poe.tasks]
dev = "uv run uvicorn app.main:app --reload --port 8080"
lint = "uv run ruff check"
format = "uv run ruff format"
test = "uv run pytest"
